commit 20200127
-content detail page update

-content Main page update

+++ PentaBeans_Project/src/main/java/com/team/board/mapper/BoardMapper.java
@@ -10,6 +10,6 @@ public interface BoardMapper {
 	public long findkey(String title);
 	public boolean fileUpload(C_fileVO fVO);
 	public ArrayList<ContentVO> getContents();
-	public ContentVO contentInfo(int cno);
-	public ArrayList<C_fileVO> fileInfo(int cno);
+	public ContentVO contentInfo(long cno);
+	public ArrayList<C_fileVO> fileInfo(long cno);
 }
